updateProfile

+++ Back/controllers/comment.controllers.js
@@ -9,7 +9,7 @@ module.exports.createComment = (req, res) => {
         res.status(400).json({ message: "Post invalide  " + err });
       } else {
         db.query(
-          `SELECT * FROM comments WHERE text LIKE %?% AND postId = ? AND posterId = ? `[
+          `SELECT * FROM comments WHERE text LIKE = ? AND postId = ? AND posterId = ? `[
             (req.body.texte, req.body.postId, req.body.posterId)
           ],
           (err, resultat) => {
@@ -26,7 +26,7 @@ module.exports.createComment = (req, res) => {
   }
 };
 
-module.exports.getAllComment = (req, res) => {
+module.exports.getComments = (req, res) => {
   try {
     db.query(`SELECT posterId, text FROM comments`, (err, result) => {
       if (!result) {
@@ -40,7 +40,7 @@ module.exports.getAllComment = (req, res) => {
   }
 };
 
-module.exports.getOneComment = (req, res) => {
+module.exports.getComment = (req, res) => {
   try {
     db.query(
       `SELECT posterId, text FROM comments WHERE id = ?`,
@@ -65,13 +65,15 @@ module.exports.modifyComment = (req, res) => {
       `UPDATE comments SET ? WHERE id = ? AND posterId = ? AND postId = ? `,
       [post, req.params.id, req.body.posterId, req.postId],
       (err, result) => {
-        if (err) {
-          return res.status(400).json({ message: "Problème put  " + err });
-        }
         if (!result) {
           return res.status(400).json({ message: "Commentaire introuvable" });
-        } else {
+        }
+        if (req.body.posterId == req.cookies.token.id) {
           res.status(200).json({ message: "Commentaire modifié" });
+        } else {
+          return res
+            .status(400)
+            .json({ message: "Problème modification  " + err });
         }
       }
     );
@@ -81,27 +83,34 @@ module.exports.modifyComment = (req, res) => {
 };
 
 module.exports.deleteComment = (req, res) => {
-  db.query(
-    `SELECT * FROM comments WHERE id = ?`,
-    req.params.id,
-    (err, resultat) => {
-      if (!resultat[0]) {
-        return res.status(400).json({ message: "Commentaire introuvable  " });
-      } else {
-        db.query(
-          `DELETE FROM comments WHERE id = ?`,
-          req.params.id,
-          (err, result) => {
-            if (err) {
-              return res
-                .status(400)
-                .json({ message: "Suppression échoué  " + err });
-            } else {
-              res.status(200).json({ message: "Commentaire supprimé" });
+  try {
+    db.query(
+      `SELECT * FROM comments WHERE id = ?`,
+      req.params.id,
+      (err, resultat) => {
+        if (!resultat[0]) {
+          return res.status(400).json({ message: "Commentaire introuvable  " });
+        }
+        if (resultat[0].posterId == req.cookies.token.id) {
+          db.query(
+            `DELETE FROM comments WHERE id = ?`,
+            req.params.id,
+            (err, result) => {
+              if (err) {
+                return res
+                  .status(400)
+                  .json({ message: "Suppression échoué  " + err });
+              } else {
+                res.status(200).json({ message: "Commentaire supprimé" });
+              }
             }
-          }
-        );
+          );
+        } else {
+          return res.status(400).json({ message: "requête non autorisé" });
+        }
       }
-    }
-  );
+    );
+  } catch (error) {
+    return res.status(400).json({ message: error });
+  }
 };
